Refactor: extract DurableEntityClient class to match .NET SDK pattern
