+++ client/src/main/java/com/microsoft/durabletask/DurableTaskGrpcClient.java
@@ -33,6 +33,7 @@ public final class DurableTaskGrpcClient extends DurableTaskClient {
     private final ManagedChannel managedSidecarChannel;
     private final TaskHubSidecarServiceBlockingStub sidecarClient;
     private final String defaultVersion;
+    private final GrpcDurableEntityClient entityClient;
 
     DurableTaskGrpcClient(DurableTaskGrpcClientBuilder builder) {
         this.dataConverter = builder.dataConverter != null ? builder.dataConverter : new JacksonDataConverter();
@@ -59,6 +60,7 @@ public final class DurableTaskGrpcClient extends DurableTaskClient {
         }
 
         this.sidecarClient = TaskHubSidecarServiceGrpc.newBlockingStub(sidecarGrpcChannel);
+        this.entityClient = new GrpcDurableEntityClient("GrpcDurableEntityClient", this.sidecarClient, this.dataConverter);
     }
 
     DurableTaskGrpcClient(int port, String defaultVersion) {
@@ -71,6 +73,7 @@ public final class DurableTaskGrpcClient extends DurableTaskClient {
                 .usePlaintext()
                 .build();
         this.sidecarClient = TaskHubSidecarServiceGrpc.newBlockingStub(this.managedSidecarChannel);
+        this.entityClient = new GrpcDurableEntityClient("GrpcDurableEntityClient", this.sidecarClient, this.dataConverter);
     }
 
     /**
@@ -388,146 +391,8 @@ public final class DurableTaskGrpcClient extends DurableTaskClient {
     // region Entity APIs
 
     @Override
-    public void signalEntity(
-            EntityInstanceId entityId,
-            String operationName,
-            @Nullable Object input,
-            @Nullable SignalEntityOptions options) {
-        Helpers.throwIfArgumentNull(entityId, "entityId");
-        Helpers.throwIfArgumentNull(operationName, "operationName");
-
-        SignalEntityRequest.Builder builder = SignalEntityRequest.newBuilder()
-                .setInstanceId(entityId.toString())
-                .setName(operationName)
-                .setRequestId(UUID.randomUUID().toString());
-
-        if (input != null) {
-            String serializedInput = this.dataConverter.serialize(input);
-            if (serializedInput != null) {
-                builder.setInput(StringValue.of(serializedInput));
-            }
-        }
-
-        if (options != null && options.getScheduledTime() != null) {
-            Timestamp ts = DataConverter.getTimestampFromInstant(options.getScheduledTime());
-            builder.setScheduledTime(ts);
-        }
-
-        this.sidecarClient.signalEntity(builder.build());
-    }
-
-    @Override
-    @Nullable
-    public EntityMetadata getEntityMetadata(EntityInstanceId entityId, boolean includeState) {
-        Helpers.throwIfArgumentNull(entityId, "entityId");
-
-        GetEntityRequest request = GetEntityRequest.newBuilder()
-                .setInstanceId(entityId.toString())
-                .setIncludeState(includeState)
-                .build();
-
-        GetEntityResponse response = this.sidecarClient.getEntity(request);
-        if (!response.getExists()) {
-            return null;
-        }
-
-        return toEntityMetadata(response.getEntity());
-    }
-
-    @Override
-    public EntityQueryResult queryEntities(EntityQuery query) {
-        Helpers.throwIfArgumentNull(query, "query");
-
-        com.microsoft.durabletask.implementation.protobuf.OrchestratorService.EntityQuery.Builder queryBuilder =
-                com.microsoft.durabletask.implementation.protobuf.OrchestratorService.EntityQuery.newBuilder();
-
-        if (query.getInstanceIdStartsWith() != null) {
-            queryBuilder.setInstanceIdStartsWith(StringValue.of(query.getInstanceIdStartsWith()));
-        }
-        if (query.getLastModifiedFrom() != null) {
-            queryBuilder.setLastModifiedFrom(DataConverter.getTimestampFromInstant(query.getLastModifiedFrom()));
-        }
-        if (query.getLastModifiedTo() != null) {
-            queryBuilder.setLastModifiedTo(DataConverter.getTimestampFromInstant(query.getLastModifiedTo()));
-        }
-        queryBuilder.setIncludeState(query.isIncludeState());
-        queryBuilder.setIncludeTransient(query.isIncludeTransient());
-        if (query.getPageSize() != null) {
-            queryBuilder.setPageSize(com.google.protobuf.Int32Value.of(query.getPageSize()));
-        }
-        if (query.getContinuationToken() != null) {
-            queryBuilder.setContinuationToken(StringValue.of(query.getContinuationToken()));
-        }
-
-        QueryEntitiesRequest request = QueryEntitiesRequest.newBuilder()
-                .setQuery(queryBuilder)
-                .build();
-
-        QueryEntitiesResponse response = this.sidecarClient.queryEntities(request);
-
-        List<EntityMetadata> entities = new ArrayList<>();
-        for (com.microsoft.durabletask.implementation.protobuf.OrchestratorService.EntityMetadata protoEntity
-                : response.getEntitiesList()) {
-            entities.add(toEntityMetadata(protoEntity));
-        }
-
-        String continuationToken = response.hasContinuationToken()
-                ? response.getContinuationToken().getValue()
-                : null;
-
-        return new EntityQueryResult(entities, continuationToken);
-    }
-
-    @Override
-    public CleanEntityStorageResult cleanEntityStorage(CleanEntityStorageRequest request) {
-        Helpers.throwIfArgumentNull(request, "request");
-
-        int totalEmptyEntitiesRemoved = 0;
-        int totalOrphanedLocksReleased = 0;
-        String continuationToken = request.getContinuationToken();
-
-        do {
-            com.microsoft.durabletask.implementation.protobuf.OrchestratorService.CleanEntityStorageRequest.Builder builder =
-                    com.microsoft.durabletask.implementation.protobuf.OrchestratorService.CleanEntityStorageRequest.newBuilder()
-                            .setRemoveEmptyEntities(request.isRemoveEmptyEntities())
-                            .setReleaseOrphanedLocks(request.isReleaseOrphanedLocks());
-
-            if (continuationToken != null) {
-                builder.setContinuationToken(StringValue.of(continuationToken));
-            }
-
-            CleanEntityStorageResponse response = this.sidecarClient.cleanEntityStorage(builder.build());
-
-            totalEmptyEntitiesRemoved += response.getEmptyEntitiesRemoved();
-            totalOrphanedLocksReleased += response.getOrphanedLocksReleased();
-
-            continuationToken = response.hasContinuationToken()
-                    ? response.getContinuationToken().getValue()
-                    : null;
-        } while (request.isContinueUntilComplete() && continuationToken != null);
-
-        return new CleanEntityStorageResult(
-                continuationToken,
-                totalEmptyEntitiesRemoved,
-                totalOrphanedLocksReleased);
-    }
-
-    private EntityMetadata toEntityMetadata(
-            com.microsoft.durabletask.implementation.protobuf.OrchestratorService.EntityMetadata protoEntity) {
-        Instant lastModifiedTime = DataConverter.getInstantFromTimestamp(protoEntity.getLastModifiedTime());
-        String lockedBy = protoEntity.hasLockedBy() ? protoEntity.getLockedBy().getValue() : null;
-        String serializedState = protoEntity.hasSerializedState()
-                ? protoEntity.getSerializedState().getValue()
-                : null;
-
-        return new EntityMetadata(
-                protoEntity.getInstanceId(),
-                lastModifiedTime,
-                protoEntity.getBacklogQueueSize(),
-                lockedBy,
-                serializedState,
-                protoEntity.hasSerializedState(),
-                this.dataConverter);
+    public DurableEntityClient getEntities() {
+        return this.entityClient;
     }
 
     // endregion
